Create method to get total cost

+++ src/TravelHop.java
@@ -33,7 +33,7 @@ public class TravelHop {
         return mode;
     }
 
-    public float getCost() {
+    public int getCost() {
         return cost;
     }
